Fix "update --direct". Disable "update + cache"
Old current server does not have a cache file. We disable
the normal update function.

The "--direct" update function will not use "lynx --dump".
So "lynx" is not needed anymore. This is because "ctan" has
a ready file to download (FILES.byname). We use "wget" to
fetch this file to local directory. After being processed,
this file is still kept in ~/.ctan/ directory, unless user
provides "--keep-only-cache" option.